partial fix for zbias property - now relatively ok, but will need more work

+++ src/services/render/ZBiasProperty.h
@@ -32,7 +32,7 @@ namespace Opde {
 
 	/** a Z-Bias property implementation using rendered property handler.
 	* Controls the rendering bias of the object (for Z-Fightning avoidance). SS2/T2 only.
-	* Uses simple unsigned float data storage. Defaults to 0.0 - No bias. Inherits always.
+	* Uses simple unsigned int data storage. Defaults to 0 - No bias. The stored bias is in bits. Inherits always.
 	*/
 	class ZBiasProperty : public RenderedProperty {
 		public:
@@ -56,7 +56,7 @@ namespace Opde {
 			void valueChanged(int oid, const std::string& field, const DVariant& value);
 
 			/// core setter method. Called from other methods to set the hasrefs value
-			void setZBias(int oid, float bias);
+			void setZBias(int oid, uint32_t bias);
 			
 			Ogre::SceneManager* mSceneMgr;
 	};
